RPM: Restart service in %posttrans, not %post
When upgrading from a 1.9 version with monolithic packaging to 1.10 or
later with per-provider subpackage, sssd-common can be upgraded (and
restarted) before the new sssd-$provider is restarted. This can lead to
a startup failure, because the sssd_be process from already upgraded
sssd-common would attempt to load a sssd_$provider.so from the
legacy sssd package.

Restarting the service in %posttrans makes sure all the packages are in
place when we restart the service.

Resolves:
https://fedorahosted.org/sssd/ticket/2399