Replace space: add some checks
This patch adds some additional checks if the option for replacing
spaces in user and group names is used.

When replacing space with the replacement character it is checked if the
name already contains the replacement character. If it does the
unmodified name is returned because in this case a revers operation
would not be possible.

For the reverse operation is it checked if the input contains both a
space and the replacement character. If this is true the unmodified name
is returned as well, because we have to assume that it is the original
name because otherwise it wouldn't contain both characters.

Additionally a shortcut if the replacement characters is a space and
tests for the new checks are added. The man page is updated accordingly.

Related to https://fedorahosted.org/sssd/ticket/1854 and
https://fedorahosted.org/sssd/ticket/2397 .

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>