refcount-tests: Do not force to run test in CK_FORK mode
There was an access to uninitialised data in old version of libcheck
when tests were executed in fork mode.

==2842== Syscall param timer_create(evp) points to uninitialised byte(s)
==2842==    at 0x977FE72: timer_create@@GLIBC_2.3.3 (timer_create.c:82)
==2842==    by 0x4E3647B: srunner_run (in /usr/lib64/libcheck.so.0.0.0)
==2842==    by 0x401616: main (refcount-tests.c:233)
==2842==  Address 0xffefff460 is on thread 1's stack

This bug is already fixed in check-0.9.14, but we want to execute tests
with valgrind on older platforms. The refcount test had hardcoded
CK_FORK in source code and therefore it could not be overwritten with
environment variable, which we use in CI script.

Reviewed-by: Nikolai Kondrashov <Nikolai.Kondrashov@redhat.com>