CI: Consider libcmocka-devel always present
Add explicit dependency on libcmocka-devel when running on any Red Hat
distros, as it turns out it exists everywhere, if only in EPEL distros,
and even though the spec file doesn't require it.

This makes the contrib/ci/run consider cmocka present on all the
supported distros, so remove the corresponding condition as well.

Reviewed-by: Michal Židek <mzidek@redhat.com>