CI: Add Valgrind suppression support
Add an empty Valgrind suppressions file, use it when invoking Valgrind.
This prepares for addition of Valgrind suppressions for current false
positives and issues that cannot be fixed, preparing for enforcing
Valgrind check.

Make Valgrind output a suppression for every error and make it output
used suppression names and counts at the end of each run. This
simplifies discovery and addition of new suppressions and removal of
unused ones.

Related to https://fedorahosted.org/sssd/ticket/2428

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>