name2sid: Check negative cache for users and groups
Since we cannot know if a SID belongs to a user or a group a lookup
should only fail if the given name is in both the negative cache for the
users and the groups.

Currently if the SID for a group called 'abc' should be looked up and
the negative cache for the users contain an entry for 'abc' the request
fails.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>