IPA: add support for new extdom plugin version
Initially the extdom plugin was only used to translate SIDs of AD user
and groups to names or POSIX IDs. On IPA clients group memberships were
resolved with the help of the PAC in the Kerberos ticket which required
that the user has logged in at least once. Home directory and the login
shell were auto generated.

The new version of the extdom plugin can return the complete list of
group memberships of a user and the list of all members of a group.
Additionally the gecos field, home directory and login shell are
returned together with an optional list of key-value pairs for arbitrary
data which is written unmodified to the cache.

Fixes https://fedorahosted.org/sssd/ticket/2159
  and https://fedorahosted.org/sssd/ticket/2041

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>