SDAP: check that connection is open before bind
Tivoli server does not return an empty response when being asked for the
rootDSE data but an error. In this case the rootDSE lookup in SSSD will
terminate the connection to the server and return a error. But since
errors except timeouts are ignored SSSD will try to continue with the
bind, but since the connection is already terminated this will fail as
well. And this will terminate the whole operation.

Make sure the connection is open before performing bind operation.

Resolves:
https://fedorahosted.org/sssd/ticket/2435

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>