sysdb_get_user_attr_with_views: add mandatory override attributes
This patch add another attribute with is needs for override processing
to the attribute list of sysdb_get_user_attr_with_views(). With two
attribute it does not seem useful to check for existence and add each of
the attributes conditionally. With this patch they are added
unconditionally if the domain has views. Additionally the attributes are
not removed in the end because it is expected that they do not cause any
harm.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>