BUILD: Use separate chown to make changing ownership to the sssd user non-fatal
When the SSSD is built in the build system using a non-root user, the
user doesn't exist in the build system and file ownership will be
maintained by the downstream packaging instead.

We need to make sure that setting the ownership to the sssd user is a
separate step from creating the directories in this case in order to
make failure to set the ownership non-fatal.

Related:
https://fedorahosted.org/sssd/ticket/2370

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>