AD: Never store case_sensitive as "true" to confdb
If case_sensitive was set 'true' for AD
backend, we ignore it and continue with AD
default (false). However we still set confdb
to whatever was set in sssd.conf for the
responders. We should store to confdb
the value that is used by the backend.

Also fixes some misleading DEBUG messages
in that code area.

Reviewed-by: Pavel Reichl <preichl@redhat.com>