BUILD: Install libsss_crypt after its dependencies
A library should not be installed before it's internal dependencies
otherwise there is an error in make distcheck.

libtool: install: error: relink `libsss_crypt.la' with the above command before
installing it

It would be sufficient just to change order of libraries in
automake variable pkglib_LTLIBRARIES, but it's better to have
internal libraries on the one place.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>