IPA: verify group memberships of trusted domain users
Depending on the state of the cache group object a freshly created or
updates user entry for a trusted domain user might already be a member
of the group or not. This cache makes sure the requested user is a
member of all groups returned from the extdom request. Special care has
to be taken to cover cross-domain group-memberships properly.

Resolves https://fedorahosted.org/sssd/ticket/2529

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>