sbus: use 'path/*' to represent a D-Bus fallback
Use 'path/*' instead of 'path*' since it better describes what
we are actually doing i.e. registering a message handler for a
subtree.

Although D-Bus fallback will invoke a message handler for both
'path' and 'path/subtree' object paths it does not make usually
sence to support the same interfaces for both parent and it
children.

This commit also renames related functions to better describe
what are they doing.

Note: the tilda in comments is used to suppress -Wcomment warning

Preparation for:
https://fedorahosted.org/sssd/ticket/2339

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>