GPO: Ignore ENOENT result from sysdb_gpo_get_gpo_result_setting()
https://fedorahosted.org/sssd/ticket/2542

If the GPO result object was missing completely, we would error out with
a fatal error code. It's more user-friendly to treat the missing object
as if the requested attribute was missing on the provider level.

Reviewed-by: Pavel Reichl <preichl@redhat.com>
(cherry picked from commit fc2cc91a5b645180e53d46436b0d08011aac8d74)