logrotate: Fix warning file size changed while zipping
Postpone compression of the previous log file to the next rotation cycle.
This only has effect when used in combination with compress. We need to use it
because we cannot tell sssd to close log files and thus sssd processes might
continue writing to the previous log file for some time.

Resolves:
https://fedorahosted.org/sssd/ticket/2547

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>