IPA: resolve IPA group-memberships for AD users
So far only for initgroups requests the IPA group memberships where
resolved for AD users and due to
6fac5e5f0c54a0f92872ce1450606cfcb577a920 those memberships are not
overridden by other request. But it turned out that the originalMemberOf
attributes related to the IPA group memberships can be overridden by
user lookups.  Since the originalMemberOf attribute is important in the
HBAC evaluation this patch makes sure that the originalMemberOf
attribute is not removed but updated during user lookups.

Related to https://fedorahosted.org/sssd/ticket/2560

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>