IPA: Use attr's dom for users, too
The 'dom' pointer points to domain of the main object being saved. In
case of group, dom points to the domain where the group resides. But
when saving members, each members might be from a different domain, so we
need to find every member's domain based on the attributes.

Also don't use Yoda style in conditions.

Reviewed-by: Sumit Bose <sbose@redhat.com>