selinux: Delete existing user mapping on empty default
https://fedorahosted.org/sssd/ticket/2587

The case of SELinux default user mapping being an empty string is valid,
it should translate into "pick the default context on the target
machine".

In case the context is empty, we need to delete the per-user mapping from
the SELinux database to make sure the default is used.

Reviewed-by: Michal Židek <mzidek@redhat.com>
Reviewed-by: Pavel Reichl <preichl@redhat.com>
(cherry picked from commit 01f78f755fde63997ccfded71fb8395569b11430)