PAM: use the logon_name as the key for the PAM initgr cache
Currently the name member of the pam_data struct is used as a key but it
can change during a request. Especially for sub-domain users the name is
changed from the short to the fully-qualified version before the cache
entry is created. As a result the cache searches are always done with
the short name while the entry was written with the fully-qualified name.

The logon_name member of the pam_data struct contains the name which was
send by the PAM client and is never changed during the request.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>