SDAP: use DN to update entry
sdap_nested_group_populate_users() has code to handle user name changes.
It updates the SYSDB_NAME attribute. This attribute is also used in the
RDN but changing the attribute in the object does not change the DN
hence the DN still contains the old name. Currently
sysdb_set_user_attr() was used to update the entry which creates the DN
based on the give name. This will fail if the name is changed for a
second time. Since the DN is already available in the search result it
is more reliable to use it directly with sysdb_set_entry_attr().

Related to https://fedorahosted.org/sssd/ticket/2591

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit 305267064a9d8c86536fcd5c92c1c9cb3e7df268)