test_expire: Use right assertion macro for standard functions
Documentation to macro assert_return_code says:
Assert that the return_code is greater than or equal to 0.

The function prints an error message to standard error and terminates the
test by calling fail() if the return code is smaller than 0. If the function
you check sets an errno if it fails you can pass it to the function and
it will be printed as part of the error message.

So in case of error we will see more verbose message.

Reviewed-by: Pavel Reichl <preichl@redhat.com>